Move linking-related code to its own file
